feat(vue-dot): move FileList into UploadWorkflow
