+++ packages/vue-dot/src/patterns/UploadWorkflow/FileList/FileList.vue
@@ -98,7 +98,7 @@
 	import { config } from './config';
 	import { FileItem } from './types';
 
-	import { customizable, Options } from '../../mixins/customizable';
+	import { customizable, Options } from '../../../mixins/customizable';
 
 	import {
 		mdiRefresh,
