+++ packages/vue-dot/src/patterns/UploadWorkflow/FileList/tests/FileList.spec.ts
@@ -4,7 +4,7 @@ import { Wrapper } from '@vue/test-utils';
 import { mountComponent } from '@/tests';
 import { html } from '@/tests/html';
 
-import FileList from '../';
+import FileList from '..';
 
 let wrapper: Wrapper<Vue>;
